src/utils/pyproject.toml
[tool.poetry]
name = "utils"
version = "0.1.0"
description = "RAG enhanced though multiple search types."
authors = ["Tyler Kendrick <tkendrick@microsoft.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
azure-search-documents = "^11.4.0"
azure-storage-blob = "^12.20.0"
streamlit = "^1.36.0"
matplotlib = "^3.9.1"
plotly = "^5.23.0"
openai = "^1.37.0"
models = { path = "../models", develop = true }
functions = { path = "../functions", develop = true }
agents = { path = "../agents", develop = true }
redis = "^5.0.8"

[tool.poetry.group.dev.dependencies]
env = "^0.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
